Threads and MVC pattern seems to be working.
I've put in a VERY temporary test, which uses the View, Controller and
Model. Since the actual MAP is not stored in the model, I am unaware how
to draw a character, using the model, relative to the MAP and not the
SCREEN...

+++ src/Model/Animal.java
@@ -15,12 +15,12 @@ public class Animal implements ICollidable {
 	}
 
 	@Override
-	public double getX() {
+	public float getX() {
 		return 0;
 	}
 
 	@Override
-	public double getY() {
+	public float getY() {
 		return 0;
 	}
